+++ CHANGELOG.md
@@ -20,6 +20,7 @@
  - Add Headers option to URL (requests) camera library
  - Add Input: MAX30155 PT100 I2C Temperature Sensor (by DFRobot)
  - Add Input: Kasa Energy Meter (Strips/Plugs)
+ - Add Inout: Atlas Scientific O2 (Oxygen Gas)
  - Add Outputs: Kasa Wifi RGB Light Bulbs
  - Add Outputs: Kasa Wifi Power Plugs
  - Add Output: Neopixel RGB LED Strip